refactor(relayer): use tx manager for all eth->gear relayers, mirror refactoring
- Removed original MessageSender in favor of a new one
- Moved proof composition code to proof_composer.rs
- Moved transaction manager code into eth_to_gear, no more dependency on
paid_token_transfers submodule
- Split tasks and tx manager into multiple functions for simplicity
