+++ relayer/src/message_relayer/common/gear/mod.rs
@@ -3,4 +3,3 @@ pub mod checkpoints_extractor;
 pub mod merkle_proof_fetcher;
 pub mod message_paid_event_extractor;
 pub mod message_queued_event_extractor;
-pub mod message_sender;
