+++ relayer/src/message_relayer/eth_to_gear/mod.rs
@@ -2,3 +2,8 @@ pub mod all_token_transfers;
 pub mod api_provider;
 pub mod manual;
 pub mod paid_token_transfers;
+
+pub mod message_sender;
+pub mod proof_composer;
+pub mod storage;
+pub mod tx_manager;
